// src/app/features/home-resume/models/skill.ts
export interface Skill {
  name: string;
  percentage: number;
  icon: string;
  /* readOnly: boolean;
  size: number;
  unit: string;
  textColor: string;
  fontSize: string;
  fontWeigth: string;
  fontFamily: string;
  valueformat: string;
  value: number;
  iconSize: number;
  iconColor: string;
  max: number;
  trackWidth: number;
  barWidth: number;
  trackColor: string;
  barColor: string;
  subText?: SubText; */
}

/* export interface SubText {
  enabled: boolean;
  fontFamily: string;
  font: string;
  fontWeight: string;
  text: string;
  color: string;
  offset: number;
} */